旅游产品和路线内容全部修改完成 完善了代码逻辑 完善了sql创建文件

+++ ruoyi-system/src/main/java/com/ruoyi/travel/mapper/ProductRouteMapper.java
@@ -26,7 +26,7 @@ public interface ProductRouteMapper
      * @param productId 产品路线ID
      * @return 产品路线
      */
-    public int countRouteNumberByProductId(Long productId);
+    public Long countRouteNumberByProductId(Long productId);
 
     /**
      * 查询产品路线列表
